third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/event-listener-breakpoints-xhr-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests event listener breakpoints on XHRs.

Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) downloadEnd (event-listener-breakpoints-xhr.js:24)
Event target: XMLHttpRequest

Call stack:
    0) downloadProgress (event-listener-breakpoints-xhr.js:32)
Event target: XMLHttpRequest

Call stack:
    0) loadCallback (event-listener-breakpoints-xhr.js:40)
Event target: XMLHttpRequest

Call stack:
    0) uploadEnd (event-listener-breakpoints-xhr.js:28)
Event target: XMLHttpRequestUpload

Call stack:
    0) uploadProgress (event-listener-breakpoints-xhr.js:36)
Event target: XMLHttpRequestUpload

Call stack:
    0) xhr.onreadystatechange (event-listener-breakpoints-xhr.js:20)
    1) sendXHR (event-listener-breakpoints-xhr.js:49)
    2) testFunction (event-listener-breakpoints-xhr.js:12)
Event target: XMLHttpRequest
